Text Renderer, Time and Movements
Now the File is more optimised than before and contains new classes like Time Class and Tex Rendering all of this is done on Day #1. The Map Class is However Still incomplete and requires lot of work to be done before submission. SDL_ttf library was also used and I used Rubik-Light.ttf font for the game. Rest Everything is Done. For Tomorrow I will start work on the Enemy and their behaviours!